+++ CHANGELOG.md
@@ -3,6 +3,77 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/jungvonmatt/contentful-ssg/compare/v1.7.2...v1.7.3) (2022-05-02)
+
+
+### Bug Fixes
+
+* **observable:** uses ReplaySubject for transforms ([89c6318](https://github.com/jungvonmatt/contentful-ssg/commit/89c6318cebbf74f4db3f126403a7ef07e8e76fbe))
+
+
+
+
+
+## [1.7.2](https://github.com/jungvonmatt/contentful-ssg/compare/v1.7.1...v1.7.2) (2022-04-28)
+
+**Note:** Version bump only for package @jungvonmatt/contentful-ssg-project
+
+
+
+
+
+## [1.7.1](https://github.com/jungvonmatt/contentful-ssg/compare/v1.7.0...v1.7.1) (2022-04-27)
+
+
+### Bug Fixes
+
+* **tests:** adds sonar reporter ([10cd59d](https://github.com/jungvonmatt/contentful-ssg/commit/10cd59d0ae6747df4010007d40e35a0d43f3bfff))
+
+
+
+
+
+# [1.7.0](https://github.com/jungvonmatt/contentful-ssg/compare/v1.6.1...v1.7.0) (2022-04-26)
+
+
+### Features
+
+* **utils:** adds waitFor util ([#39](https://github.com/jungvonmatt/contentful-ssg/issues/39)) ([0840ecd](https://github.com/jungvonmatt/contentful-ssg/commit/0840ecde90a3e08cebde144dafabb720c2e2124f))
+
+
+
+
+
+## [1.6.1](https://github.com/jungvonmatt/contentful-ssg/compare/v1.6.0...v1.6.1) (2022-04-05)
+
+**Note:** Version bump only for package @jungvonmatt/contentful-ssg-project
+
+
+
+
+
+# [1.6.0](https://github.com/jungvonmatt/contentful-ssg/compare/v1.5.2...v1.6.0) (2022-03-28)
+
+
+### Features
+
+* **asset-sizes:** allow function passed in sizes array ([cbf3b28](https://github.com/jungvonmatt/contentful-ssg/commit/cbf3b2863f51925ff16b0fa869bf444b3bc1959a))
+
+
+
+
+
+## [1.5.2](https://github.com/jungvonmatt/contentful-ssg/compare/v1.5.1...v1.5.2) (2022-03-23)
+
+
+### Bug Fixes
+
+* **menu:** use submenu as default value for fieldIdMenu ([e7a8b55](https://github.com/jungvonmatt/contentful-ssg/commit/e7a8b55a61cb3d5284561b318d0bf5b2a92a835e))
+
+
+
+
+
 ## [1.5.1](https://github.com/jungvonmatt/contentful-ssg/compare/v1.5.0...v1.5.1) (2022-02-23)