+++ packages/cssg-plugin-hugo/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.7.3](https://github.com/jungvonmatt/contentful-ssg/compare/v1.7.2...v1.7.3) (2022-05-02)
+
+**Note:** Version bump only for package @jungvonmatt/cssg-plugin-hugo
+
+
+
+
+
+## [1.7.2](https://github.com/jungvonmatt/contentful-ssg/compare/v1.7.1...v1.7.2) (2022-04-28)
+
+**Note:** Version bump only for package @jungvonmatt/cssg-plugin-hugo
+
+
+
+
+
+## [1.7.1](https://github.com/jungvonmatt/contentful-ssg/compare/v1.7.0...v1.7.1) (2022-04-27)
+
+**Note:** Version bump only for package @jungvonmatt/cssg-plugin-hugo
+
+
+
+
+
+# [1.7.0](https://github.com/jungvonmatt/contentful-ssg/compare/v1.6.1...v1.7.0) (2022-04-26)
+
+**Note:** Version bump only for package @jungvonmatt/cssg-plugin-hugo
+
+
+
+
+
+## [1.6.1](https://github.com/jungvonmatt/contentful-ssg/compare/v1.6.0...v1.6.1) (2022-04-05)
+
+**Note:** Version bump only for package @jungvonmatt/cssg-plugin-hugo
+
+
+
+
+
+## [1.5.2](https://github.com/jungvonmatt/contentful-ssg/compare/v1.5.1...v1.5.2) (2022-03-23)
+
+
+### Bug Fixes
+
+* **menu:** use submenu as default value for fieldIdMenu ([e7a8b55](https://github.com/jungvonmatt/contentful-ssg/commit/e7a8b55a61cb3d5284561b318d0bf5b2a92a835e))
+
+
+
+
+
 ## [1.5.1](https://github.com/jungvonmatt/contentful-ssg/compare/v1.5.0...v1.5.1) (2022-02-23)